Create UpdateListProc
This is one of the store proc's I wrote for "list" app I am writing.  The two main tables are "Lists" and "Items" and obviously the point is to be able to create "Lists" that have "Items".  The relational design means that I need a pivot table so that I don't have redundant items in the Items table, which I have called "ListItems".  When you create a list, you may add items to that list.  This is a pretty easy operation:  Insert to the Lists and insert to the ListItems.  But if you Update the list, you need to be able insert, update and delete, depending on certain settings.

I decided that using a Table Value Parameter was the best way to handle multiple ListItem requests.  From C# I send an Enumerable List of ListItems and that is received as Tvp.  After the List is updated, I have three queries: insert (based on if there is no primary key id value), update (achieved by doing an inner join between the ListItem table and the Tvp) and delete (if quantity is 0, there is no point in pointing to an Item).